+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 * Add ability to add fields from SQL Database via in GenericAdder
+* Prometheus Exporter now exports also processor specific metrics
 * Add `--version` cli argument to print the current logprep version, as well as the configuration
 version if found
 
@@ -11,12 +12,21 @@ version if found
 * Automatically release logprep on pypi
 * Configure abstract dependencies for pypi releases
 * Refactor domain resolver
+* Refactor `processor_stats` to `metrics`. Metrics are now collected in separate dataclasses
 
 ### Bugfixes
 
 * Fix processor initialization in auto rule tester
 * Fix generation of RST-Docs
 
+### Breaking 
+
+* Metrics refactoring: 
+  * The json output format of the previously known status_logger has changed
+  * The configuration key word is now `metrics` instead of `status_logger`
+  * The configuration for the time measurement is now part of the metrics configuration
+  * The metrics tracking still includes values about how many warnings and errors happened, but 
+  not of what type. For that the regular logprep logging should be consolidated.
 
 ## v2.0.1